restructure of metricqueueservice, working kafka rewrite of query queueing
