+++ ArgusClient/src/main/java/com/salesforce/dva/argus/client/MetricProcessor.java
@@ -1,12 +1,9 @@
 package com.salesforce.dva.argus.client;
 
 import com.salesforce.dva.argus.entity.AsyncBatchedMetricQuery;
-import com.salesforce.dva.argus.service.MetricQueueService;
-import com.salesforce.dva.argus.service.MetricService;
+import com.salesforce.dva.argus.service.BatchService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Created by cguan on 6/6/16.
@@ -20,22 +17,17 @@ public class MetricProcessor implements Runnable {
 
     //~ Instance fields ******************************************************************************************************************************
 
-    private final MetricQueueService metricQueueService;
-    private final MetricService metricService;
-    private final AtomicInteger jobCounter;
+    private final BatchService batchService;
 
     //~ Constructors *********************************************************************************************************************************
 
     /**
      * Creates a new MetricProcessor object.
      *
-     * @param  metricQueueService The async metric queueing service to use.
-     * @param  metricService The metric service to use.
+     * @param  batchService     The batch service to use. Cannot be null.
      */
-    MetricProcessor(MetricQueueService metricQueueService, MetricService metricService, AtomicInteger jobCounter) {
-        this.metricQueueService = metricQueueService;
-        this.metricService = metricService;
-        this.jobCounter = jobCounter;
+    MetricProcessor(BatchService batchService) {
+        this.batchService = batchService;
     }
 
     //~ Methods **************************************************************************************************************************************
@@ -44,7 +36,7 @@ public class MetricProcessor implements Runnable {
     public void run() {
         while (!Thread.currentThread().isInterrupted()) {
             try {
-                AsyncBatchedMetricQuery query = metricQueueService.dequeueAndProcess(1);
+                AsyncBatchedMetricQuery query = batchService.executeNextQuery(10000);
                 Thread.sleep(POLL_INTERVAL_MS);
             } catch (InterruptedException ex) {
                 LOGGER.info("Execution was interrupted.");
