+++ ArgusCore/src/main/java/com/salesforce/dva/argus/service/metric/AsyncMetricService.java
@@ -2,17 +2,12 @@ package com.salesforce.dva.argus.service.metric;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-import com.salesforce.dva.argus.entity.AsyncBatchedMetricQuery;
 import com.salesforce.dva.argus.entity.BatchMetricQuery;
 import com.salesforce.dva.argus.entity.Metric;
-import com.salesforce.dva.argus.inject.SLF4JTypeListener;
 import com.salesforce.dva.argus.service.BatchService;
-import com.salesforce.dva.argus.service.CacheService;
-import com.salesforce.dva.argus.service.MetricQueueService;
 import com.salesforce.dva.argus.service.MonitorService;
 import com.salesforce.dva.argus.service.tsdb.MetricQuery;
 import com.salesforce.dva.argus.system.SystemConfiguration;
-import org.slf4j.Logger;
 
 import java.util.List;
 
@@ -24,21 +19,16 @@ import java.util.List;
 public class AsyncMetricService extends DefaultMetricService {
 
     //~ Instance fields ******************************************************************************************************************************
-    @SLF4JTypeListener.InjectLogger
-    private Logger _logger;
-    private final MetricQueueService _metricQueueService;
-    private final CacheService _cacheService;
+
     private final BatchService _batchService;
 
     //~ Constructors *********************************************************************************************************************************
 
     @Inject
     protected AsyncMetricService(MonitorService monitorService, Provider<MetricReader<Metric>> metricsprovider,
-                                 Provider<MetricReader<MetricQuery>> queryprovider, MetricQueueService metricQueueService,
-                                 CacheService cacheService, BatchService batchService, SystemConfiguration config) {
+                                 Provider<MetricReader<MetricQuery>> queryprovider,
+                                 BatchService batchService, SystemConfiguration config) {
         super(monitorService, metricsprovider, queryprovider, config);
-        _metricQueueService = metricQueueService;
-        _cacheService = cacheService;
         _batchService = batchService;
     }
 
@@ -46,14 +36,10 @@ public class AsyncMetricService extends DefaultMetricService {
 
     @Override
     public String getAsyncMetrics(List<String> expressions, long offset, int ttl, String ownerName) {
-        _logger.debug("ASYNCMETRICSERVICE GOOD LOGGER");
-        int priority = 1;
-        BatchMetricQuery batchQuery = new BatchMetricQuery(expressions, offset, priority, ttl, ownerName);
-        for (AsyncBatchedMetricQuery query: batchQuery.getQueries()) {
-            _metricQueueService.enqueue(query, 1);
-        }
-        _batchService.updateBatch(batchQuery);
-        return batchQuery.getBatchId();
+        BatchMetricQuery batch= new BatchMetricQuery(expressions, offset, ttl, ownerName);
+        _batchService.enqueueBatch(batch);
+        _batchService.updateBatch(batch);
+        return batch.getBatchId();
     }
 }
 /* Copyright (c) 2016, Salesforce.com, Inc.  All rights reserved. */
