+++ ArgusCore/src/main/java/com/salesforce/dva/argus/system/SystemInitializer.java
@@ -43,13 +43,12 @@ import com.salesforce.dva.argus.inject.SLF4JTypeListener;
 import com.salesforce.dva.argus.service.*;
 import com.salesforce.dva.argus.service.annotation.DefaultAnnotationService;
 import com.salesforce.dva.argus.service.audit.DefaultAuditService;
-import com.salesforce.dva.argus.service.batch.CachedBasedBatchService;
+import com.salesforce.dva.argus.service.batch.DefaultBatchService;
 import com.salesforce.dva.argus.service.collect.DefaultCollectionService;
 import com.salesforce.dva.argus.service.history.DefaultHistoryService;
 import com.salesforce.dva.argus.service.jpa.*;
 import com.salesforce.dva.argus.service.management.DefaultManagementService;
 import com.salesforce.dva.argus.service.metric.AsyncMetricService;
-import com.salesforce.dva.argus.service.metric.queue.CacheBasedMetricQueueService;
 import com.salesforce.dva.argus.service.monitor.DefaultMonitorService;
 import com.salesforce.dva.argus.service.schema.DefaultDiscoveryService;
 import com.salesforce.dva.argus.service.tsdb.CachedTSDBService;
@@ -217,8 +216,7 @@ final class SystemInitializer extends AbstractModule {
         bindConcreteClass(DefaultDashboardService.class, DashboardService.class);
         bindConcreteClass(DefaultCollectionService.class, CollectionService.class);
         bindConcreteClass(AsyncMetricService.class, MetricService.class);
-        bindConcreteClass(CacheBasedMetricQueueService.class, MetricQueueService.class);
-        bindConcreteClass(CachedBasedBatchService.class, BatchService.class);
+        bindConcreteClass(DefaultBatchService.class, BatchService.class);
         bindConcreteClass(DefaultGlobalInterlockService.class, GlobalInterlockService.class);
         bindConcreteClass(DefaultMonitorService.class, MonitorService.class);
         bindConcreteClass(DefaultWardenService.class, WardenService.class);
